+++ resources/clients/ubirch-default-realm/environments/environment.prod.ts
@@ -11,6 +11,7 @@ const keycloakConfig: KeycloakConfig = {
 
 export const environment = {
   production: true,
+  envName: 'prod',
   serverUrl: '@@UBIRCH_TR_ENV_SERVURL@@', // 'http://localhost:8081',
   apiPrefix: '@@UBIRCH_TR_API_PREF@@', // '/ubirch-web-ui/api/v1/',
   keyServiceServerUrl: '@@UBIRCH_TR_ENV_KEYSERVICE_SERVURL@@', // 'https://key.dev.ubirch.com',