+++ resources/clients/ubirch-default-realm/environments/environment.ts
@@ -6,17 +6,18 @@
 import {KeycloakConfig} from 'keycloak-angular';
 
 const keycloakConfig: KeycloakConfig = {
-  url: 'http://localhost:8080/auth',
+  url: 'https://id.dev.ubirch.com/auth',
   realm: 'ubirch-default-realm',
   clientId: 'ubirch-2.0-user-access',
   credentials: {
-    secret: 'df168843-a441-4721-9f2b-588cb3c208c5'
+    secret: 'KEYCLOAK-CLIENT-SECRET'
   }
 };
 
 export const environment = {
   production: false,
-  serverUrl: 'http://localhost:8081',
+  envName: 'dev',
+  serverUrl: 'https://api.console.dev.ubirch.com',
   apiPrefix: '/ubirch-web-ui/api/v1/',
   keyServiceServerUrl: 'https://key.dev.ubirch.com',
   keyServiceApiPrefix: '/api/keyService/v1/',