+++ run-local.sh
@@ -1,13 +1,25 @@
 #!/bin/bash
 
 client=$1
+staging=$2
 
 if [[ -z "$client" ]]
 then
-    client=test-realm
-    echo "secure app by realm 'test-realm'"
+    client=ubirch-default-realm
+    echo "secure app by realm 'ubirch-default-realm'"
 fi
 
 ./preprocess4client.sh $client
 
-ionic serve --port=9101
+./preprocess-replace-secrets.sh $client
+
+
+if [[ -z "$staging" ]]
+then
+    echo "run app on default stage (dev)"
+    ionic serve --port=9101
+else
+    echo "run app on stage '$staging'"
+    ionic serve --port=9101 --configuration=$staging
+fi
+