+++ src/app/app.component.ts
@@ -52,7 +52,7 @@ export class AppComponent {
     this.platform.ready().then(() => {
       this.statusBar.styleDefault();
       this.splashScreen.hide();
-      console.log('version 0.3.3');
+      console.log('version 0.4.0');
     });
   }
 }