+++ src/app/services/device.service.ts
@@ -1,16 +1,17 @@
-import { Injectable } from '@angular/core';
+import {Injectable} from '@angular/core';
 import {BehaviorSubject, Observable, of, throwError} from 'rxjs';
 import {Device} from '../models/device';
 import {FormGroup} from '@angular/forms';
 import {UbirchWebUIUtilsService} from '../utils/ubirch-web-uiutils.service';
 import {environment} from '../../environments/environment';
 import {HttpClient} from '@angular/common/http';
-import {catchError, map, tap} from 'rxjs/operators';
+import {catchError, map, switchMap, tap} from 'rxjs/operators';
 import {DeviceTypeService} from './device-type.service';
 import {CreateDevicesFormData} from '../tabs/devices/devices-list-page/popups/new-device-popup/new-device-popup.component';
 import {DevicesListWrapper} from '../models/devices-list-wrapper';
 import {UserService} from './user.service';
 import {isArray} from 'util';
+import {DeviceState, TIME_RANGES} from '../models/device-state';
 
 @Injectable({
   providedIn: 'root'
@@ -18,7 +19,9 @@ import {isArray} from 'util';
 export class DeviceService {
     url = environment.serverUrl + environment.apiPrefix;
     devicesUrl = this.url + 'devices';  // URL to web api to access devices
+    deviceStateUrl = this.url + 'devices/state';  // URL to web api to access the states of requested devices
     searchUrl = this.devicesUrl + '/search';  // URL to web api to search devices by hwDeviceId or description (substrings)
+
     private currentDevice: Device;
     private behaviorSubject = new BehaviorSubject<Device>(this.currentDevice);
     public observableCurrentDevice: Observable<Device> = this.behaviorSubject.asObservable();
@@ -32,23 +35,63 @@ export class DeviceService {
   ) { }
 
     /**
-     * load a couple of deviceStubs for the list of devices; pagination included
+     * load a couple of deviceStubs for the list of devices; pagination included, states fetched after list has been responded
      * @param pageNum number of the page for pagination
      * @param pageSize how many devices are maximal paged
      */
     public reloadDeviceStubs(pageNum?: number, pageSize?: number): Observable<DevicesListWrapper> {
-        const url = UbirchWebUIUtilsService.addParamsToURL(
+        const url = UbirchWebUIUtilsService.addParamsToThingsListPaginationURL(
             this.devicesUrl,
             pageNum,
             pageSize);
 
         return this.http.get<DevicesListWrapper[]>(url).pipe(
-            map(listWrapper => new DevicesListWrapper(listWrapper)),
-            tap(listWrapper => {
-                if (listWrapper && listWrapper.numberOfDevices) {
-                    this.userService.setNumberOfDevices(listWrapper.numberOfDevices);
+            switchMap(data => {
+              const listWrapper = new DevicesListWrapper(data);
+              if (listWrapper) {
+                if (listWrapper.numberOfDevices) {
+                  this.userService.setNumberOfDevices(listWrapper.numberOfDevices);
                 }
-        }));
+                if (listWrapper.devices) {
+                  return this.getDeviceStates(listWrapper.devices.map(device => device.hwDeviceId)).pipe(
+                    map(states => {
+                      listWrapper.setDeviceStates(states);
+                      return listWrapper;
+                    }
+                  ));
+                } else {
+                  return of(listWrapper);
+                }
+              }
+            }
+        ));
+    }
+
+    /**
+     * Get states of devices
+     * @param deviceIds array of deviceId for which the states should be requested for
+     * @param from Date Object (optional) if given this is the end of the request range of time, states are request for; default: Date.now()
+     * @param periode Date Object (optional) if given this is range of time, states are request for; default: 1 day
+     */
+    public getDeviceStates(deviceIds: string[], from?: Date, to?: Date): Observable<DeviceState[]> {
+      if (deviceIds && isArray(deviceIds) && deviceIds.length > 0) {
+        const endDate = to ? to : new Date();
+        const startDate = from ? from : new Date(endDate.getTime() - TIME_RANGES.DAY);
+
+        const url = UbirchWebUIUtilsService.addParamsToURL(
+          this.deviceStateUrl,
+          [ startDate.toISOString(),
+            endDate.toISOString()
+          ]
+        );
+
+        return this.http.post<DeviceState[]>(url, deviceIds.join()).pipe(
+          map(jsonStates =>
+            jsonStates.map(
+              jsonState => new DeviceState(jsonState))));
+      } else {
+        throw new Error(`getDeviceStates call without any deviceIds`);
+      }
     }
 
     /**
@@ -165,10 +208,10 @@ export class DeviceService {
 
     }
 
-  storeUnsavedChangesOfDevice(val: any): boolean {
-    // TODO: how will we store changes of device before saving it?
-    return true;
-  }
+    storeUnsavedChangesOfDevice(val: any): boolean {
+      // TODO: how will we store changes of device before saving it?
+      return true;
+    }
 
     private data2Devices(data: CreateDevicesFormData): Device[] {
         const devicesArray: Device[] = [];
@@ -185,7 +228,5 @@ export class DeviceService {
             });
         }
         return devicesArray;
-
     }
-
 }